+++ packages/script/src/script.ts
@@ -7,6 +7,7 @@ import type {
   TransactionResultReturnDataReceipt,
   TransactionResultRevertReceipt,
   TransactionResultScriptResultReceipt,
+  TransactionResult,
 } from '@fuel-ts/providers';
 import { ReceiptType } from '@fuel-ts/transactions';
 
@@ -112,7 +113,10 @@ export class Script<TData = void, TResult = void> {
       const decoded = this.scriptResultDecoder(scriptResult);
       return decoded;
     } catch (error) {
-      throw new ScriptResultDecoderError(callResult, (error as Error).message);
+      throw new ScriptResultDecoderError(
+        callResult as TransactionResult<'failure'>,
+        (error as Error).message
+      );
     }
   }
 }