+++ packages/fuels/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Change Log
 
+## 0.12.0
+
+### Patch Changes
+
+- Updated dependencies [[`b45c3cd`](https://github.com/FuelLabs/fuels-ts/commit/b45c3cd9d834e11f7ccc358bb93103b937c2a9fe), [`f3dcd27`](https://github.com/FuelLabs/fuels-ts/commit/f3dcd272dc5a237c6a8ce235a542ad804039f13b), [`400607d`](https://github.com/FuelLabs/fuels-ts/commit/400607d3e957d9cb6a2925d35b7b10393d6f15cc)]:
+  - @fuel-ts/contract@0.12.0
+  - @fuel-ts/abi-coder@0.12.0
+  - @fuel-ts/address@0.12.0
+  - @fuel-ts/interfaces@0.12.0
+  - @fuel-ts/predicate@0.12.0
+  - @fuel-ts/providers@0.12.0
+  - @fuel-ts/signer@0.12.0
+  - @fuel-ts/testcases@0.12.0
+  - @fuel-ts/wallet@0.12.0
+  - @fuel-ts/wallet-manager@0.12.0
+  - @fuel-ts/transactions@0.12.0
+  - @fuel-ts/hasher@0.12.0
+  - @fuel-ts/hdwallet@0.12.0
+  - @fuel-ts/constants@0.12.0
+  - @fuel-ts/keystore@0.12.0
+  - @fuel-ts/math@0.12.0
+  - @fuel-ts/merkle@0.12.0
+  - @fuel-ts/merkle-shared@0.12.0
+  - @fuel-ts/merklesum@0.12.0
+  - @fuel-ts/mnemonic@0.12.0
+  - @fuel-ts/sparsemerkle@0.12.0
+  - @fuel-ts/wordlists@0.12.0
+
 ## 0.11.0
 
 ### Patch Changes